refactor(threads): useThreads 拆 memoryCapture/autoNaming/completionEmail 三域（3385→2446）
