+++ src/features/threads/hooks/useThreadAutoNaming.ts
@@ -1,0 +1,447 @@
+import { useCallback } from "react";
+import type { Dispatch, MutableRefObject } from "react";
+import type { DebugEntry } from "../../../types";
+import {
+  generateThreadTitle,
+  listThreadTitles,
+  resumeThread,
+} from "../../../services/tauri";
+import { buildItemsFromThread } from "../../../utils/threadItems";
+import i18n from "../../../i18n";
+import {
+  makeCustomNameKey,
+  saveCustomName,
+  type CustomNamesMap,
+} from "../utils/threadStorage";
+import type { ThreadAction, ThreadState } from "./useThreadsReducer";
+
+const AUTO_TITLE_REQUEST_TIMEOUT_MS = 8_000;
+const AUTO_TITLE_MAX_ATTEMPTS = 2;
+const AUTO_TITLE_PENDING_STALE_MS = 20_000;
+
+type UseThreadAutoNamingOptions = {
+  customNamesRef: MutableRefObject<CustomNamesMap>;
+  dispatch: Dispatch<ThreadAction>;
+  onDebug?: (entry: DebugEntry) => void;
+  clearAutoTitlePending: (workspaceId: string, threadId: string) => void;
+  getAutoTitlePendingStartedAt: (
+    workspaceId: string,
+    threadId: string,
+  ) => number | null;
+  isAutoTitlePending: (workspaceId: string, threadId: string) => boolean;
+  markAutoTitlePending: (workspaceId: string, threadId: string) => void;
+  state: ThreadState;
+};
+
+export function useThreadAutoNaming({
+  customNamesRef,
+  dispatch,
+  onDebug,
+  clearAutoTitlePending,
+  getAutoTitlePendingStartedAt,
+  isAutoTitlePending,
+  markAutoTitlePending,
+  state,
+}: UseThreadAutoNamingOptions) {
+  const autoNameThread = useCallback(
+    async (
+      workspaceId: string,
+      threadId: string,
+      sourceText: string,
+      options?: { force?: boolean; clearPendingOnSkip?: boolean },
+    ): Promise<string | null> => {
+      const key = makeCustomNameKey(workspaceId, threadId);
+      const hasCustomName = Boolean(customNamesRef.current[key]);
+      if (hasCustomName && !options?.force) {
+        onDebug?.({
+          id: `${Date.now()}-thread-title-skip-custom`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "thread/title skipped",
+          payload: { workspaceId, threadId, reason: "has-custom-name" },
+        });
+        if (options?.clearPendingOnSkip) {
+          clearAutoTitlePending(workspaceId, threadId);
+        }
+        return null;
+      }
+
+      const message = sourceText.trim();
+      if (!message) {
+        onDebug?.({
+          id: `${Date.now()}-thread-title-skip-empty`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "thread/title skipped",
+          payload: { workspaceId, threadId, reason: "empty-source-text" },
+        });
+        if (options?.clearPendingOnSkip) {
+          clearAutoTitlePending(workspaceId, threadId);
+        }
+        return null;
+      }
+
+      const pendingStartedAt = getAutoTitlePendingStartedAt(
+        workspaceId,
+        threadId,
+      );
+      if (pendingStartedAt) {
+        const pendingAgeMs = Date.now() - pendingStartedAt;
+        if (pendingAgeMs >= AUTO_TITLE_PENDING_STALE_MS) {
+          onDebug?.({
+            id: `${Date.now()}-thread-title-pending-timeout-reset`,
+            timestamp: Date.now(),
+            source: "client",
+            label: "thread/title pending reset",
+            payload: {
+              workspaceId,
+              threadId,
+              pendingStartedAt,
+              pendingAgeMs,
+              reason: "timeout",
+            },
+          });
+          clearAutoTitlePending(workspaceId, threadId);
+        } else {
+          onDebug?.({
+            id: `${Date.now()}-thread-title-skip-pending`,
+            timestamp: Date.now(),
+            source: "client",
+            label: "thread/title skipped",
+            payload: {
+              workspaceId,
+              threadId,
+              reason: "already-pending",
+              pendingStartedAt,
+              pendingAgeMs,
+            },
+          });
+          return null;
+        }
+      }
+
+      if (isAutoTitlePending(workspaceId, threadId)) {
+        onDebug?.({
+          id: `${Date.now()}-thread-title-skip-pending-after-reset`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "thread/title skipped",
+          payload: {
+            workspaceId,
+            threadId,
+            reason: "already-pending-after-reset",
+          },
+        });
+        return null;
+      }
+
+      markAutoTitlePending(workspaceId, threadId);
+      const markAt = getAutoTitlePendingStartedAt(workspaceId, threadId);
+      onDebug?.({
+        id: `${Date.now()}-thread-title-generate-start`,
+        timestamp: Date.now(),
+        source: "client",
+        label: "thread/title generate",
+        payload: {
+          workspaceId,
+          threadId,
+          force: Boolean(options?.force),
+          pendingStartedAt: markAt,
+        },
+      });
+
+      try {
+        const applyGeneratedTitle = (
+          title: string,
+          source: "generated" | "recovered",
+        ) => {
+          saveCustomName(workspaceId, threadId, title);
+          const nextKey = makeCustomNameKey(workspaceId, threadId);
+          customNamesRef.current[nextKey] = title;
+          dispatch({
+            type: "setThreadName",
+            workspaceId,
+            threadId,
+            name: title,
+          });
+          onDebug?.({
+            id: `${Date.now()}-thread-title-${source}-success`,
+            timestamp: Date.now(),
+            source: "server",
+            label:
+              source === "generated"
+                ? "thread/title generated"
+                : "thread/title recovered",
+            payload: { workspaceId, threadId, title, source },
+          });
+          return title;
+        };
+
+        const generateWithTimeout = async (
+          preferredLanguage: "zh" | "en",
+        ): Promise<string> =>
+          await new Promise((resolve, reject) => {
+            const timeoutId = setTimeout(() => {
+              reject(new Error("auto-title-timeout"));
+            }, AUTO_TITLE_REQUEST_TIMEOUT_MS);
+
+            void generateThreadTitle(
+              workspaceId,
+              threadId,
+              message,
+              preferredLanguage,
+            ).then(
+              (value) => {
+                clearTimeout(timeoutId);
+                resolve(value);
+              },
+              (error) => {
+                clearTimeout(timeoutId);
+                reject(error);
+              },
+            );
+          });
+
+        const language = i18n.language.toLowerCase().startsWith("zh")
+          ? "zh"
+          : "en";
+        for (
+          let attempt = 1;
+          attempt <= AUTO_TITLE_MAX_ATTEMPTS;
+          attempt += 1
+        ) {
+          const attemptStartedAt = Date.now();
+          try {
+            onDebug?.({
+              id: `${Date.now()}-thread-title-attempt-start`,
+              timestamp: Date.now(),
+              source: "client",
+              label: "thread/title attempt",
+              payload: {
+                workspaceId,
+                threadId,
+                attempt,
+                maxAttempts: AUTO_TITLE_MAX_ATTEMPTS,
+                timeoutMs: AUTO_TITLE_REQUEST_TIMEOUT_MS,
+                language,
+              },
+            });
+
+            const generated = await generateWithTimeout(language);
+            const title = generated.trim();
+            if (!title) {
+              throw new Error("empty-generated-title");
+            }
+            return applyGeneratedTitle(title, "generated");
+          } catch (error) {
+            const errorMessage =
+              error instanceof Error ? error.message : String(error);
+            const isTimeout = errorMessage.includes("auto-title-timeout");
+            const elapsedMs = Date.now() - attemptStartedAt;
+
+            onDebug?.({
+              id: `${Date.now()}-thread-title-attempt-failed`,
+              timestamp: Date.now(),
+              source: isTimeout ? "client" : "error",
+              label: "thread/title attempt failed",
+              payload: {
+                workspaceId,
+                threadId,
+                attempt,
+                maxAttempts: AUTO_TITLE_MAX_ATTEMPTS,
+                isTimeout,
+                elapsedMs,
+                error: errorMessage,
+              },
+            });
+
+            if (isTimeout) {
+              await new Promise((resolve) => setTimeout(resolve, 1200));
+            }
+
+            try {
+              const mappedTitles = await listThreadTitles(workspaceId);
+              const recovered = mappedTitles[threadId]?.trim();
+              if (recovered) {
+                return applyGeneratedTitle(recovered, "recovered");
+              }
+            } catch (recoveryError) {
+              onDebug?.({
+                id: `${Date.now()}-thread-title-recovery-check-error`,
+                timestamp: Date.now(),
+                source: "error",
+                label: "thread/title recovery error",
+                payload:
+                  recoveryError instanceof Error
+                    ? recoveryError.message
+                    : String(recoveryError),
+              });
+            }
+
+            if (attempt < AUTO_TITLE_MAX_ATTEMPTS) {
+              onDebug?.({
+                id: `${Date.now()}-thread-title-retry`,
+                timestamp: Date.now(),
+                source: "client",
+                label: "thread/title retry",
+                payload: {
+                  workspaceId,
+                  threadId,
+                  nextAttempt: attempt + 1,
+                  maxAttempts: AUTO_TITLE_MAX_ATTEMPTS,
+                },
+              });
+              continue;
+            }
+
+            onDebug?.({
+              id: `${Date.now()}-thread-title-generate-error`,
+              timestamp: Date.now(),
+              source: "error",
+              label: "thread/title generate error",
+              payload: {
+                workspaceId,
+                threadId,
+                attempt,
+                maxAttempts: AUTO_TITLE_MAX_ATTEMPTS,
+                error: errorMessage,
+              },
+            });
+            return null;
+          }
+        }
+
+        return null;
+      } catch (error) {
+        onDebug?.({
+          id: `${Date.now()}-thread-title-generate-error`,
+          timestamp: Date.now(),
+          source: "error",
+          label: "thread/title generate error",
+          payload: {
+            workspaceId,
+            threadId,
+            error: error instanceof Error ? error.message : String(error),
+          },
+        });
+        return null;
+      } finally {
+        clearAutoTitlePending(workspaceId, threadId);
+      }
+    },
+    [
+      clearAutoTitlePending,
+      customNamesRef,
+      dispatch,
+      getAutoTitlePendingStartedAt,
+      isAutoTitlePending,
+      markAutoTitlePending,
+      onDebug,
+    ],
+  );
+
+  const triggerAutoThreadTitle = useCallback(
+    async (
+      workspaceId: string,
+      threadId: string,
+      options?: { force?: boolean },
+    ) => {
+      const items = state.itemsByThread[threadId] ?? [];
+      const userMessage = items.find(
+        (item) => item.kind === "message" && item.role === "user",
+      );
+      let text =
+        userMessage && userMessage.kind === "message" ? userMessage.text : "";
+
+      if (!text.trim() && !threadId.startsWith("claude:")) {
+        try {
+          const response = (await resumeThread(
+            workspaceId,
+            threadId,
+          )) as Record<string, unknown> | null;
+          const result = (response?.result ?? response) as Record<
+            string,
+            unknown
+          > | null;
+          const thread = (result?.thread ?? response?.thread ?? null) as Record<
+            string,
+            unknown
+          > | null;
+          if (thread) {
+            const loadedItems = buildItemsFromThread(thread);
+            const loadedFirstUserMessage = loadedItems.find(
+              (item) => item.kind === "message" && item.role === "user",
+            );
+            if (
+              loadedFirstUserMessage &&
+              loadedFirstUserMessage.kind === "message" &&
+              loadedFirstUserMessage.text.trim()
+            ) {
+              text = loadedFirstUserMessage.text;
+              onDebug?.({
+                id: `${Date.now()}-thread-title-manual-source-resume`,
+                timestamp: Date.now(),
+                source: "client",
+                label: "thread/title manual source",
+                payload: { workspaceId, threadId, source: "thread/resume" },
+              });
+            }
+          }
+        } catch (error) {
+          onDebug?.({
+            id: `${Date.now()}-thread-title-manual-source-resume-error`,
+            timestamp: Date.now(),
+            source: "error",
+            label: "thread/title manual source error",
+            payload: error instanceof Error ? error.message : String(error),
+          });
+        }
+      }
+
+      if (!text.trim()) {
+        const fallbackName =
+          state.threadsByWorkspace[workspaceId]
+            ?.find((thread) => thread.id === threadId)
+            ?.name?.trim() ?? "";
+        if (fallbackName && !/^agent\s+\d+$/i.test(fallbackName)) {
+          text = fallbackName;
+          onDebug?.({
+            id: `${Date.now()}-thread-title-manual-source-name`,
+            timestamp: Date.now(),
+            source: "client",
+            label: "thread/title manual source",
+            payload: { workspaceId, threadId, source: "thread/name" },
+          });
+        }
+      }
+
+      if (!text.trim()) {
+        onDebug?.({
+          id: `${Date.now()}-thread-title-manual-missing-source`,
+          timestamp: Date.now(),
+          source: "client",
+          label: "thread/title manual skipped",
+          payload: { workspaceId, threadId, reason: "no-user-message-found" },
+        });
+      }
+      const generated = await autoNameThread(workspaceId, threadId, text, {
+        force: options?.force ?? true,
+        clearPendingOnSkip: true,
+      });
+      return generated;
+    },
+    [autoNameThread, onDebug, state.itemsByThread, state.threadsByWorkspace],
+  );
+
+  const isThreadAutoNaming = useCallback(
+    (workspaceId: string, threadId: string) =>
+      isAutoTitlePending(workspaceId, threadId),
+    [isAutoTitlePending],
+  );
+
+  return {
+    autoNameThread,
+    triggerAutoThreadTitle,
+    isThreadAutoNaming,
+  };
+}
